backends/pyproject.toml
[tool.poetry]
name = "final-pjt"
version = "0.1.0"
description = ""
authors = ["JAEHWAN"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.16"
djangorestframework = "^3.15.2"
dj-rest-auth = "^7.0.0"
django-allauth = "^65.2.0"
requests = "^2.32.3"
psycopg = "^3.2.3"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"
tzdata = "^2024.2"
pyspark = "^3.5.3"
openai = "^1.55.0"
python-dotenv = "^1.0.1"
selenium = "^4.26.1"
tqdm = "^4.67.0"
lxml = "^5.3.0"
langchain = "^0.3.7"
langchain-openai = "^0.2.9"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
